+++ src/presentation/controllers/signup/signup-protocols.ts
@@ -1,0 +1,4 @@
+export * from '../../protocols'
+export * from '../../protocols/email-validation'
+export * from '../../../domain/usecases/add-account-use-case'
+export * from '../../../domain/models/account'
