+++ src/presentation/protocols/index.ts
@@ -1,3 +1,2 @@
 export * from './https'
 export * from './controllers'
-export * from './email-validation'
